+++ cargo-insta/src/cli.rs
@@ -103,6 +103,9 @@ pub struct TestCommand {
     /// Follow up with review.
     #[structopt(long)]
     pub review: bool,
+    /// Accept all snapshots after test.
+    #[structopt(long, conflicts_with = "review")]
+    pub accept: bool,
     /// Do not reject pending snapshots before run.
     #[structopt(long)]
     pub keep_pending: bool,
@@ -173,25 +176,32 @@ fn handle_color(color: &Option<String>) -> Result<(), Box<dyn Error>> {
     Ok(())
 }
 
-#[allow(clippy::type_complexity)]
-fn handle_target_args(
-    target_args: &TargetArgs,
-) -> Result<(PathBuf, Option<Vec<Package>>, Vec<&str>), Box<dyn Error>> {
+struct LocationInfo<'a> {
+    workspace_root: PathBuf,
+    packages: Option<Vec<Package>>,
+    exts: Vec<&'a str>,
+}
+
+fn handle_target_args(target_args: &TargetArgs) -> Result<LocationInfo<'_>, Box<dyn Error>> {
     let mut exts: Vec<&str> = target_args.extensions.iter().map(|x| x.as_str()).collect();
     if exts.is_empty() {
         exts.push("snap");
     }
     match target_args.workspace_root {
-        Some(ref root) => Ok((root.clone(), None, exts)),
+        Some(ref root) => Ok(LocationInfo {
+            workspace_root: root.clone(),
+            packages: None,
+            exts,
+        }),
         None => {
             let metadata =
                 get_package_metadata(target_args.manifest_path.as_ref().map(|x| x.as_path()))?;
             let packages = find_packages(&metadata, target_args.all)?;
-            Ok((
-                metadata.workspace_root().to_path_buf(),
-                Some(packages),
+            Ok(LocationInfo {
+                workspace_root: metadata.workspace_root().to_path_buf(),
+                packages: Some(packages),
                 exts,
-            ))
+            })
         }
     }
 }
@@ -200,7 +210,11 @@ fn process_snapshots(cmd: &ProcessCommand, op: Option<Operation>) -> Result<(), 
     let term = Term::stdout();
     let mut snapshot_containers = vec![];
 
-    let (workspace_root, packages, exts) = handle_target_args(&cmd.target_args)?;
+    let LocationInfo {
+        workspace_root,
+        packages,
+        exts,
+    } = handle_target_args(&cmd.target_args)?;
 
     match packages {
         Some(ref packages) => {
@@ -353,13 +367,17 @@ fn test_run(cmd: &TestCommand) -> Result<(), Box<dyn Error>> {
         return Err(QuietExit(1).into());
     }
 
-    if cmd.review {
+    if cmd.review || cmd.accept {
         process_snapshots(
             &ProcessCommand {
                 target_args: cmd.target_args.clone(),
                 quiet: false,
             },
-            None,
+            if cmd.accept {
+                Some(Operation::Accept)
+            } else {
+                None
+            },
         )?
     }