+++ src/macros.rs
@@ -332,10 +332,12 @@ macro_rules! assert_snapshot_matches {
 
 /// Settings configuration macro.
 ///
-/// This macro lets you bind some settings temporarily.  Currently
-/// only `sort_maps` is supported:
+/// This macro lets you bind some settings temporarily.  The first argument
+/// takes key value pairs that should be set, the second is the block to
+/// execute.  All settings can be set (`sort_maps => value` maps roughly
+/// to `set_sort_maps(value)`).
 ///
-/// ```rust,ignore
+/// ```rust
 /// insta::with_settings!({sort_maps => true}, {
 ///     // run snapshot test here
 /// });
@@ -343,7 +345,7 @@ macro_rules! assert_snapshot_matches {
 #[macro_export]
 macro_rules! with_settings {
     ({$($k:ident => $v:expr),*}, $body:block) => {{
-        let mut settings = crate::Settings::new();
+        let mut settings = $crate::Settings::new();
         $(
             settings._private_inner_mut().$k = $v.into();
         )*