+++ CHANGELOG.md
@@ -6,3 +6,7 @@
 * make RON support optional.
 * added support for sorting maps before serialization.
 * added settings support.
+* added support for overriding the snapshot path.
+* added an `--accept` command to `cargo insta test`
+* correctly handle nested macros that might contain inline snapshots.
+* use thread name as snapshot name for inline snapshots.